requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
bidict==0.22.1
charset-normalizer==3.3.1
click==8.1.7
fastapi==0.104.0
frozenlist==1.4.0
gunicorn==21.2.0
h11==0.14.0
hiredis==2.2.3
httptools==0.6.1
idna==3.4
multidict==6.0.4
netifaces==0.10.6
packaging==23.2
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
python-engineio==4.8.0
python-socketio==5.10.0
PyYAML==6.0.1
redis==5.0.1
simple-websocket==1.0.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
ujson==5.8.0
uvicorn==0.23.2
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
wsproto==1.2.0
yarl==1.9.2
